split reward module to two modules: reward and gringotts
