+++ node/runtime/Cargo.toml
@@ -42,6 +42,7 @@ kton = { package = "evo-kton", path = '../../srml/kton', default-features = fals
 staking = { package = "evo-staking", path = "../../srml/staking", default-features = false}
 aura = { package = "srml-aura",  path = "../../srml/aura", default-features = false }
 reward = { package = "evo-reward", path = "../../srml/reward", default-features = false}
+gringotts = { package = "evo-gringotts", path = "../../srml/gringotts", default-features = false}
 
 [features]
 default = ["std"]
@@ -81,4 +82,5 @@ std = [
 	"offchain-primitives/std",
 	"kton/std",
 	"reward/std",
+	"gringotts/std",
 ]
