+++ srml/reward/src/lib.rs
@@ -1,62 +1,22 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
-use parity_codec::{Decode, Encode};
-use primitives::traits::{
-    CheckedSub, Zero, Bounded
-};
-
-use rstd::prelude::*;
+use primitives::traits::Zero;
 use rstd::{result, convert::{ TryInto, TryFrom}};
-use srml_support::{decl_event, decl_module, decl_storage, StorageMap, StorageValue, ensure};
+use srml_support::{decl_event, decl_module, decl_storage, StorageMap, StorageValue};
 use srml_support::traits::{
-    Currency, ExistenceRequirement, Imbalance, LockableCurrency, LockIdentifier,
-    WithdrawReason, WithdrawReasons,
+    Currency, ExistenceRequirement, Imbalance, LockableCurrency, WithdrawReason
 };
-use substrate_primitives::U256;
 use system::ensure_signed;
 use dsupport::traits::OnAccountBalanceChanged;
 use dsupport::traits::OnMinted;
 
-#[cfg(feature = "std")]
-use runtime_io::with_storage;
-
 mod mock;
 mod tests;
-
-const DEPOSIT_ID: LockIdentifier = *b"lockkton";
-
-#[derive(Encode, Decode, Clone, PartialEq, Eq)]
-#[cfg_attr(feature = "std", derive(Debug))]
-pub struct BalanceLock<Balance, BlockNumber> {
-    pub id: LockIdentifier,
-    pub amount: Balance,
-    pub until: BlockNumber,
-    pub reasons: WithdrawReasons,
-}
-
-
-#[derive(PartialEq, Eq, PartialOrd, Ord, Clone, Encode, Decode, Default)]
-#[cfg_attr(feature = "std", derive(Debug))]
-pub struct IndividualDeposit<Currency, Balance, Moment> {
-    pub month: u32,
-    pub start_at: Moment,
-    pub value: Currency,
-    pub balance: Balance,
-    pub claimed: bool,
-}
-
-#[derive(PartialEq, Eq, PartialOrd, Ord, Clone, Encode, Decode, Default)]
-#[cfg_attr(feature = "std", derive(Debug))]
-pub struct Deposit<Currency, Balance, Moment> {
-    pub total: Currency,
-    pub deposit_list: Vec<IndividualDeposit<Currency, Balance, Moment>>,
-}
 
 type KtonBalanceOf<T> = <<T as Trait>::Kton as Currency<<T as system::Trait>::AccountId>>::Balance;
 type RingBalanceOf<T> = <<T as Trait>::Ring as Currency<<T as system::Trait>::AccountId>>::Balance;
 
 pub type RingNegativeImbalanceOf<T> = <<T as Trait>::Ring as Currency<<T as system::Trait>::AccountId>>::NegativeImbalance;
-pub type RingPositiveImbalanceOf<T> = <<T as Trait>::Ring as Currency<<T as system::Trait>::AccountId>>::PositiveImbalance;
 
 pub trait Trait: timestamp::Trait {
     type Kton: Currency<Self::AccountId>;
@@ -68,13 +28,8 @@ pub trait Trait: timestamp::Trait {
 decl_event!(
     pub enum Event<T> where
     < T as system::Trait>::AccountId,
-    Balance = KtonBalanceOf<T>,
     Currency = RingBalanceOf<T>,
-    Moment = < T as timestamp::Trait>::Moment,
     {
-        /// lock ring for getting kton
-        NewDeposit(Moment, AccountId, Balance, Currency),
-
         /// Claim Reward
         RewardClaim(AccountId, Currency),
     }
@@ -82,8 +37,6 @@ decl_event!(
 
 decl_storage! {
     trait Store for Module<T: Trait> as Reward {
-        pub DepositLedger get(deposit_ledger): map T::AccountId => Option<Deposit<RingBalanceOf<T>, KtonBalanceOf<T>, T::Moment>>;
-
         pub SysAcc get(sys_acc) config(): T::AccountId;
 
         // reward you can get per kton
@@ -96,59 +49,12 @@ decl_storage! {
         /// same to balance in ring
         /// TODO: it's ugly, ready for hacking
         pub SysRevenuePot get(system_revenue): map T::AccountId => RingBalanceOf<T>;
-
-
     }
 }
 
 decl_module! {
     pub struct Module<T: Trait> for enum Call where origin: T::Origin {
         fn deposit_event<T>() = default;
-
-        pub fn deposit(origin, value: RingBalanceOf<T>, months: u32) {
-            ensure!(!months.is_zero() && months <= 36, "months must be at least 1");
-            let transactor = ensure_signed(origin)?;
-            if <DepositLedger<T>>::exists(&transactor) {
-                return Err("Already deposited.");
-            }
-
-            let free_currency = T::Ring::free_balance(&transactor);
-            let value = value.min(free_currency);
-
-            let now = <timestamp::Module<T>>::now();
-
-            let kton_return = Self::compute_kton_balance(months, value).unwrap();
-
-            let individual_deposit = IndividualDeposit {month: months, start_at: now.clone(), value: value, balance: kton_return, claimed: false};
-            let deposit = Deposit {total: value, deposit_list: vec![individual_deposit]};
-
-            Self::update_deposit(&transactor, &deposit);
-
-            let _positive_imbalance = T::Kton::deposit_creating(&transactor, kton_return);
-            Self::deposit_event(RawEvent::NewDeposit(now, transactor, kton_return, value));
-        }
-
-        fn deposit_extra(origin, additional_value: RingBalanceOf<T>, months: u32) {
-            ensure!(!months.is_zero() && months <= 36, "months must be at least 1");
-            let transactor = ensure_signed(origin)?;
-            let mut deposit = Self::deposit_ledger(&transactor).ok_or("Use fn deposit instead.")?;
-
-            let now = <timestamp::Module<T>>::now();
-            let free_currency = T::Ring::free_balance(&transactor);
-
-            if let Some(extra) = free_currency.checked_sub(&deposit.total) {
-                let extra = extra.min(additional_value);
-                deposit.total += extra;
-
-                let kton_return = Self::compute_kton_balance(months, extra).unwrap();
-                let individual_deposit = IndividualDeposit {month: months, start_at: now.clone(), value: extra.clone(), balance: kton_return, claimed: false};
-                deposit.deposit_list.push(individual_deposit);
-                Self::update_deposit(&transactor, &deposit);
-
-                let _positive_imbalance = T::Kton::deposit_creating(&transactor, kton_return);
-                Self::deposit_event(RawEvent::NewDeposit(now, transactor, kton_return, extra));
-            }
-        }
 
         pub fn claim_reward(origin) {
             let transactor = ensure_signed(origin)?;
@@ -159,38 +65,24 @@ decl_module! {
                 Self::deposit_event(RawEvent::RewardClaim(transactor, value_can_withdraw));
             }
         }
+
     }
 }
 
 impl<T: Trait> Module<T> {
-    fn update_deposit(who: &T::AccountId, deposit: &Deposit<RingBalanceOf<T>, KtonBalanceOf<T>, T::Moment>) {
-        T::Ring::set_lock(
-            DEPOSIT_ID,
-            &who,
-            deposit.total,
-            // u32::max_value().into(),
-            T::BlockNumber::max_value(),
-            WithdrawReasons::all(),
-        );
-        <DepositLedger<T>>::insert(who, deposit);
-    }
 
-    fn compute_kton_balance(months: u32, value: RingBalanceOf<T>) -> Option<KtonBalanceOf<T>> {
-        let months = months as u64;
-        let value = value.try_into().unwrap_or_default() as u64;
+    pub fn reward_to_pot(value: RingBalanceOf<T>) {
+        let sys_acc = Self::sys_acc();
+        let _positive = T::Ring::deposit_creating(&sys_acc, value);
 
-        if !months.is_zero() {
-            let no = U256::from(67_u128).pow(U256::from(months));
-            let de = U256::from(66_u128).pow(U256::from(months));
+        // update reward-per-share
+        let total_issuance: u64 = T::Kton::total_issuance().try_into().unwrap_or_default() as u64;
+        //TODO: if kton total_issuance is super high
+        // this will be zero
+        let additional_reward_per_share = value / total_issuance.try_into().unwrap_or_default();
+        <RewardPerShare<T>>::mutate(|r| *r += additional_reward_per_share);
 
-            let quotient = no / de;
-            let remainder = no % de;
-            let res = U256::from(value) * (U256::from(1000) * (quotient - 1) + U256::from(1000) * remainder / de) / U256::from(1970000);
-
-            Some(res.as_u64().try_into().unwrap_or_default())
-        } else {
-            None
-        }
+        <SysRevenuePot<T>>::insert(&sys_acc, Self::system_revenue(&sys_acc) + value);
     }
 
     fn convert_to_paid_out(value: KtonBalanceOf<T>) -> RingBalanceOf<T> {
@@ -210,21 +102,6 @@ impl<T: Trait> Module<T> {
         }
     }
 
-    fn reward_to_pot(value: RingBalanceOf<T>) {
-        let sys_acc = Self::sys_acc();
-        let _positive = T::Ring::deposit_creating(&sys_acc, value);
-
-        // update reward-per-share
-        let total_issuance: u64 = T::Kton::total_issuance().try_into().unwrap_or_default() as u64;
-        //TODO: if kton total_issuance is super high
-        // this will be zero
-        let additional_reward_per_share = value / total_issuance.try_into().unwrap_or_default();
-        <RewardPerShare<T>>::mutate(|r| *r += additional_reward_per_share);
-
-        <SysRevenuePot<T>>::insert(&sys_acc, Self::system_revenue(&sys_acc) + value);
-    }
-
-    // PUB IMMUTABLE
     fn reward_can_withdraw(who: &T::AccountId) -> RingBalanceOf<T> {
         let free_balance = T::Kton::free_balance(who);
         let max_should_withdraw = Self::convert_to_paid_out(free_balance);
@@ -235,7 +112,6 @@ impl<T: Trait> Module<T> {
         } else {
             u64::try_from(should_withdraw).unwrap_or_default().try_into().unwrap_or_default()
         }
-
     }
 
     /// pay system fee with reward
